requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.1
requests==2.32.4
tencentcloud-sdk-python==3.0.1443
urllib3==2.5.0
Werkzeug==3.1.3
cos-python-sdk-v5>=1.9.0
python-dotenv>=1.0.0
